refactor to drop LoopPattern class and added comment to LoopPatternNode class
